+++ src/BooleanDates.php
@@ -46,7 +46,7 @@ trait BooleanDates
         }
 
         if ($this->hasBooleanDate($key)) {
-            return $this->attributes[$this->getBooleanDateField($key)] !== null;
+            return $this->getBooleanDate($key);
         }
 
         return parent::getAttribute($key);
@@ -82,33 +82,37 @@ trait BooleanDates
     /**
      * @return array
      */
-    public function getBooleanDateAttributes() : array
+    protected function getBooleanDateAttributes() : array
     {
         return array_intersect_key(
-            $this->attributes,
+            $this->getAttributes(),
             array_flip($this->getBooleanDates())
         );
     }
 
     /**
-     * @param string $key
+     * @param mixed $key
      *
      * @return bool
      */
-    public function hasBooleanDate(string $key) : bool
+    protected function getBooleanDate($key) : bool
     {
-        return array_key_exists($key, $this->getBooleanDates());
+        return parent::getAttribute($this->getBooleanDateField($key)) !== null;
     }
 
     /**
      * @param string $key
      * @param mixed $value
      */
-    public function setBooleanDate(string $key, $value) : void
+    protected function setBooleanDate(string $key, $value) : void
     {
         // Only update the timestamp if the value is true and if it's not yet set
         // or if the value is false and we need to unset the field.
         if (! $value || ($value && $this->currentBooleanDateFieldValueIsNotYetSet($key))) {
+            // Set the value directly on the attributes array, don't use
+            // setAttribute, and don't receive $200. (\) (°,,,°) (/)
+            // This allows us to format and set the datetime ourselves,
+            // and makes using the $dates field optional.
             $this->attributes[$this->getBooleanDateField($key)] = $this->getNewBooleanDateValue($value);
         }
     }
@@ -118,21 +122,31 @@ trait BooleanDates
      *
      * @return bool
      */
-    public function currentBooleanDateFieldValueIsNotYetSet(string $key) : bool
+    protected function hasBooleanDate(string $key) : bool
     {
-        if (! array_key_exists($this->getBooleanDateField($key), $this->attributes)) {
-            return true;
-        }
-
-        return $this->attributes[$this->getBooleanDateField($key)] === null;
+        return array_key_exists($key, $this->getBooleanDates());
     }
 
     /**
-     * @param string $key
+     * @param mixed $key
+     *
+     * @return bool
+     */
+    protected function currentBooleanDateFieldValueIsNotYetSet($key) : bool
+    {
+        if (! array_key_exists($this->getBooleanDateField($key), $this->getAttributes())) {
+            return true;
+        }
+
+        return parent::getAttribute($this->getBooleanDateField($key)) === null;
+    }
+
+    /**
+     * @param mixed $key
      *
      * @return string
      */
-    public function getBooleanDateField(string $key) : string
+    protected function getBooleanDateField($key) : string
     {
         return $this->booleanDates[$key];
     }
@@ -142,7 +156,7 @@ trait BooleanDates
      *
      * @return string|null
      */
-    public function getNewBooleanDateValue($value) : ?string
+    protected function getNewBooleanDateValue($value) : ?string
     {
         return $value ? $this->fromDateTime(Carbon::now()) : null;
     }